Validate functions in a separate file.
